Add support for process groups
These occur with numprocs > 0 and show up in the supervisor status as:
myproc:myproc-00                 RUNNING    pid 4153, uptime 0:19:13
myproc:myproc-01                 RUNNING    pid 4153, uptime 0:19:13
myproc:myproc-03                 RUNNING    pid 4153, uptime 0:19:13

Update the status gatherer to take this into account.  If the process is
in a mix of states the state is returned as MIXED.  This will allow
start, stop and restart actions to still complete successfully and bring all the
processes to the desired state.